requirements.txt
chromadb==1.5.1
fastapi==0.133.1
httpx==0.28.1
numpy==2.4.2
pandas==3.0.1
pyarrow>=15.0
Pillow==12.1.1
pydantic==2.12.5
pydantic_settings==2.13.1
python-dotenv==1.2.1
Requests==2.32.5
scikit_learn==1.8.0
sentence_transformers==5.2.0
torch==2.10.0
transformers==4.57.6
protobuf>=4.25,<6
git+https://github.com/openai/CLIP.git
